RING: Add IEEE float conversion functions  for 360 view data

+++ engines/ring/base/stream.h
@@ -58,9 +58,9 @@ private:
 		int32 field_4;
 		int32 field_8;
 		int32 field_C;
-		int32 field_10;
-		int32 field_14;
-		int32 field_18;
+		float field_10;
+		float field_14;
+		float field_18;
 		int32 field_1C;
 		int32 field_20;
 		int32 field_24;
